shaders done, commenting in progress

+++ core/src/com/mundoviventem/render/ShaderParams.java
@@ -3,11 +3,16 @@ package com.mundoviventem.render;
 
 import java.util.ArrayList;
 
+/**
+ * Data class for encapsulating a shader string alias and it's uniforms as well as in the case of a global shader
+ * its lifetime
+ */
+
 public class ShaderParams {
 
     private String shader;
     private ArrayList<CustomUniform> uniforms = new ArrayList<>();
-    private Integer lifetime = 0;
+    private Double lifetime = 0.0;
 
     public ShaderParams(String shaderString){
         shader = shaderString;
@@ -18,7 +23,15 @@ public class ShaderParams {
         uniforms = uniformList;
     }
 
-    public ShaderParams(String shaderString, ArrayList<CustomUniform> uniformList, Integer shaderLifetime){
+    /**
+     * Creates a object possessing the alias string, a list of uniforms and its lifetime.
+     * Uniforms as well as lifetime are optional.
+     * @param shaderString
+     * @param uniformList
+     * @param shaderLifetime
+     */
+
+    public ShaderParams(String shaderString, ArrayList<CustomUniform> uniformList, Double shaderLifetime){
         shader = shaderString;
         uniforms = uniformList;
         lifetime = shaderLifetime;
@@ -31,7 +44,7 @@ public class ShaderParams {
     public ArrayList<CustomUniform> getUniforms(){
         return uniforms;
     }
-    public Integer getLifetime(){
+    public Double getLifetime(){
         return lifetime;
     }
 }
